transmission: drop from repository
It's quite unfortunate to see a great tool such as transmission go
completely south with a single release. It seems the developers worked
really hard to crap on the source (by switching from C to C++, adding
unnecessary new dependencies etc.) and the build system (by switching
from autoconf to CMake). People, if you are really inclined for some
reason to switch to a new build system, at least switch to Meson and not
CMake. It's $CURRENT_YEAR for heaven's sake.